backend/requirements.txt
django==4.2.5
djangorestframework==3.14.0
psycopg2-binary==2.9.8
python-dotenv==1.0.0
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.5
pytest==7.0
python-decouple==3.8
Pillow==8.4.0
pytest-django==4.8.0
pika==1.3.2
django-filter==23.5
google-auth>=2.0.0,<3.0.0